+++ src/Symfony2/Bundle/DoctrineBundle/Command/CreateDatabaseDoctrineCommand.php
@@ -14,9 +14,9 @@
 
 namespace Symfony2\Bundle\DoctrineBundle\Command;
 
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Output\OutputInterface;
+use Symfony2\Component\Console\Input\InputOption;
+use Symfony2\Component\Console\Input\InputInterface;
+use Symfony2\Component\Console\Output\OutputInterface;
 use Doctrine\DBAL\DriverManager;
 
 /**